requirements.txt
Flask==1.1.2
Flask-Mail==0.9.1
Werkzeug==1.0.1
python-dotenv==0.14.0
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
